move condition watcher to pkg/gateway
