+++ plugins/alerting/pkg/alerting/interfaces.go
@@ -8,7 +8,3 @@ type WatcherHooks[T any] interface {
 	RegisterEvent(eventType func(T) bool, hooks ...func(context.Context, T) error)
 	HandleEvent(T)
 }
-
-type InternalConditionWatcher interface {
-	WatchEvents()
-}
