+++ plugins/alerting/pkg/alerting/system.go
@@ -2,6 +2,7 @@ package alerting
 
 import (
 	"context"
+	"github.com/rancher/opni/pkg/gateway"
 	"os"
 	"time"
 
@@ -94,7 +95,7 @@ func (p *Plugin) UseWatchers(client managementv1.ManagementClient) {
 		func() { p.watchGlobalClusterHealthStatus(client, NewAgentStream()) },
 		func() { p.watchCortexClusterStatus() }
 
-	p.globalWatchers = NewSimpleInternalConditionWatcher(
+	p.globalWatchers = gateway.NewSimpleInternalConditionWatcher(
 		clusterCrud,
 		clusterHealthStatus,
 		cortexBackendStatus,
